@@ -11,22 +11,22 @@
   "devDependencies": {
     "@babel/runtime": "7.21.0",
     "@types/node": "18.11.18",
-    "@typescript-eslint/eslint-plugin": "^5.56.0",
-    "@typescript-eslint/parser": "^5.56.0",
+    "@typescript-eslint/eslint-plugin": "^7.14.1",
+    "@typescript-eslint/parser": "^7.14.1",
     "buffer": "^5.5.0",
-    "eslint": "^8.26.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-no-only-tests": "^3.1.0",
-    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-react": "^7.34.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "next": "14.1.4",
     "nx": "19.3.2",
     "prettier": "2.8.6",
     "process": "^0.11.10",
     "ts-node": "10.9.1",
-    "typescript": "5.4.2",
+    "typescript": "5.5.2",
     "vitest": "0.28.3"
   }
 }
