@@ -1,8 +1,8 @@
 import * as React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import Styleguide from './components/Styleguide';
 
-ReactDOM.render(<Styleguide />, document.getElementById(`root`));
+ReactDOM.createRoot(container()).render(<Styleguide />);
 
 // lol should SSR this instead ¯\_(ツ)_/¯
 if (window.location.hash) {
@@ -11,4 +11,12 @@ if (window.location.hash) {
   if (el) {
     setTimeout(() => el.scrollIntoView(), 750);
   }
+}
+
+function container(): ReactDOM.Container {
+  const app = document.getElementById(`root`);
+  if (!app) {
+    throw new Error(`No element with id 'root' found in document`);
+  }
+  return app;
 }
